refactor: types.d.ts -> /lib/dtos.d.ts 변경
